[Schema & Reader] split schema and data
