+++ core/src/main/java/cn/edu/tsinghua/iotdb/benchmark/extern/SchemaWriter.java
@@ -2,13 +2,12 @@ package cn.edu.tsinghua.iotdb.benchmark.extern;
 
 import cn.edu.tsinghua.iotdb.benchmark.conf.Config;
 import cn.edu.tsinghua.iotdb.benchmark.conf.ConfigDescriptor;
-import cn.edu.tsinghua.iotdb.benchmark.entity.Batch;
 import cn.edu.tsinghua.iotdb.benchmark.schema.MetaDataSchema;
 import cn.edu.tsinghua.iotdb.benchmark.schema.schemaImpl.DeviceSchema;
 
 import java.util.List;
 
-public abstract class BasicWriter {
+public abstract class SchemaWriter {
   protected static final Config config = ConfigDescriptor.getInstance().getConfig();
   protected static final MetaDataSchema metaDataSchema = MetaDataSchema.getInstance();
 
@@ -17,8 +16,9 @@ public abstract class BasicWriter {
    *
    * @return
    */
-  public static BasicWriter getBasicWriter() {
-    return new CSVWriter();
+  public static SchemaWriter getBasicWriter() {
+    // TODO change according to config
+    return new CSVSchemaWriter();
   }
 
   /**
@@ -28,13 +28,4 @@ public abstract class BasicWriter {
    * @return
    */
   public abstract boolean writeSchema(List<DeviceSchema> deviceSchemaList);
-
-  /**
-   * Write Batch to the file
-   *
-   * @param batch
-   * @param insertLoopIndex loop index of batch
-   * @return
-   */
-  public abstract boolean writeBatch(Batch batch, long insertLoopIndex);
 }
